+++ cmd/pubsub/subscription/main.go
@@ -17,142 +17,26 @@ limitations under the License.
 package main
 
 import (
-	"errors"
 	"flag"
-	"log"
-	"time"
 
-	"cloud.google.com/go/pubsub"
-
-	"cloud.google.com/go/compute/metadata"
+	"github.com/GoogleCloudPlatform/cloud-run-events/pkg/pubsub/operations"
 	"github.com/kelseyhightower/envconfig"
+	"github.com/knative/pkg/logging"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 	"golang.org/x/net/context"
 )
-
-type envConfig struct {
-	// Environment variable containing project id.
-	Project string `envconfig:"PROJECT_ID"`
-
-	// Action is the operation the job should run.
-	// Options: [exists, create, delete]
-	Action string `envconfig:"ACTION" required:"true"`
-
-	// Topic is the environment variable containing the PubSub Topic being
-	// subscribed to's name. In the form that is unique within the project.
-	// E.g. 'laconia', not 'projects/my-gcp-project/topics/laconia'.
-	Topic string `envconfig:"PUBSUB_TOPIC_ID" required:"true"`
-
-	// Subscription is the environment variable containing the name of the
-	// subscription to use.
-	Subscription string `envconfig:"PUBSUB_SUBSCRIPTION_ID" required:"true"`
-}
-
-// TODO: the job could output the resolved projectID.
 
 func main() {
 	flag.Parse()
 
-	ctx := context.Background()
-	logCfg := zap.NewProductionConfig() // TODO: to replace with a dynamically updating logger.
-	logCfg.EncoderConfig.EncodeTime = zapcore.ISO8601TimeEncoder
-	logger, err := logCfg.Build()
-	if err != nil {
-		log.Fatalf("Unable to create logger: %v", err)
-	}
+	logger, _ := logging.NewLogger("", "INFO") // TODO: use logging config map.
+	defer logger.Sync()
 
-	var env envConfig
-	if err := envconfig.Process("", &env); err != nil {
+	ctx := logging.WithLogger(context.Background(), logger)
+
+	ops := operations.SubscriptionOps{}
+	if err := envconfig.Process("", &ops); err != nil {
 		logger.Fatal("Failed to process env var", zap.Error(err))
 	}
-
-	if env.Project == "" {
-		project, err := metadata.ProjectID()
-		if err != nil {
-			logger.Fatal("failed to find project id. ", zap.Error(err))
-		}
-		env.Project = project
-	}
-
-	logger = logger.With(
-		zap.String("action", env.Action),
-		zap.String("project", env.Project),
-		zap.String("topic", env.Topic),
-		zap.String("subscription", env.Subscription),
-	)
-
-	logger.Info("Pub/Sub Subscription Job.")
-
-	client, err := pubsub.NewClient(ctx, env.Project)
-	if err != nil {
-		logger.Fatal("Failed to create Pub/Sub client.", zap.Error(err))
-	}
-
-	// Load the subscription.
-	sub := client.Subscription(env.Subscription)
-	exists, err := sub.Exists(ctx)
-	if err != nil {
-		logger.Fatal("Failed to verify topic exists.", zap.Error(err))
-	}
-
-	switch env.Action {
-	case "exists":
-		// If subscription doesn't exist, that is an error.
-		if !exists {
-			logger.Fatal("Subscription does not exist.")
-		}
-		logger.Info("Previously created.")
-
-	case "create":
-		// If topic doesn't exist, create it.
-		if !exists {
-			// Load the topic.
-			topic, err := getTopic(ctx, client, env.Topic)
-			if err != nil {
-				logger.Fatal("Failed to get topic.", zap.Error(err))
-			}
-			// Create a new subscription to the previous topic with the given name.
-			sub, err = client.CreateSubscription(ctx, env.Subscription, pubsub.SubscriptionConfig{
-				Topic:             topic,
-				AckDeadline:       30 * time.Second,
-				RetentionDuration: 25 * time.Hour,
-			})
-			if err != nil {
-				logger.Fatal("Failed to create subscription.", zap.Error(err))
-			}
-			logger.Info("Successfully created.")
-		} else {
-			// TODO: here is where we could update config.
-			logger.Info("Previously created.")
-		}
-
-	case "delete":
-		if exists {
-			if err := sub.Delete(ctx); err != nil {
-				logger.Fatal("Failed to delete subscription.", zap.Error(err))
-			}
-			logger.Info("Successfully deleted.")
-		} else {
-			logger.Info("Previously deleted.")
-		}
-
-	default:
-		logger.Fatal("unknown action value.")
-	}
-
-	logger.Info("Done.")
-}
-
-func getTopic(ctx context.Context, client *pubsub.Client, topicID string) (*pubsub.Topic, error) {
-	// Load the topic.
-	topic := client.Topic(topicID)
-	ok, err := topic.Exists(ctx)
-	if err != nil {
-		return nil, err
-	}
-	if ok {
-		return topic, err
-	}
-	return nil, errors.New("topic does not exist")
+	ops.Run(ctx)
 }
